use NotificationStatuses in test_utils
so that we don't make this mistake of forgetting about pending again
