+++ tests/functional/staging_and_prod/notify_api/test_notify_api_email.py
@@ -4,10 +4,9 @@ from config import config
 from tests.postman import (
     send_notification_via_api,
     get_notification_by_id_via_api,
-    NotificationStatuses
 )
 
-from tests.test_utils import assert_notification_body, recordtime
+from tests.test_utils import assert_notification_body, recordtime, NotificationStatuses
 
 
 @recordtime
