+++ tests/postman.py
@@ -48,13 +48,6 @@ def send_notification_via_csv(upload_csv_page, message_type, seeded=False):
     return notification_id
 
 
-class NotificationStatuses:
-    PENDING_VIRUS_CHECK = 'pending-virus-check'
-    RECEIVED = {'received'}
-    DELIVERED = {'delivered', 'temporary-failure', 'permanent-failure'}
-    SENT = RECEIVED | DELIVERED | {'sending', 'pending'}
-
-
 def get_notification_by_id_via_api(client, notification_id, expected_statuses):
     try:
         resp = client.get_notification_by_id(notification_id)
